refactor(supabase): consolidate realtime initialization and table constants
