+++ src/modules/diet/day-diet/application/usecases/dayState.ts
@@ -1,10 +1,10 @@
 import { createEffect } from 'solid-js'
 
-import { initializeDayDietRealtime } from '~/modules/diet/day-diet/application/realtime'
 import { dayCacheStore } from '~/modules/diet/day-diet/infrastructure/signals/dayCacheStore'
 import { dayChangeStore } from '~/modules/diet/day-diet/infrastructure/signals/dayChangeStore'
 import { initializeDayEffects } from '~/modules/diet/day-diet/infrastructure/signals/dayEffects'
 import { dayStateStore } from '~/modules/diet/day-diet/infrastructure/signals/dayStateStore'
+import { initializeDayDietRealtime } from '~/modules/diet/day-diet/infrastructure/supabase/realtime'
 import { createDebug } from '~/shared/utils/createDebug'
 
 const debug = createDebug()
