+++ src/modules/diet/day-diet/infrastructure/supabase/constants.ts
@@ -1,0 +1,1 @@
+export const SUPABASE_TABLE_DAYS = 'days'
