--- src/modules/diet/macro-profile/application/realtime.ts
@@ -1,11 +1,0 @@
-import { initializeMacroProfileRealtime as initializeMacroProfileRealtimeInfra } from '~/modules/diet/macro-profile/infrastructure/supabase/realtime'
-
-let initialized = false
-
-export function initializeMacroProfileRealtime() {
-  if (initialized) {
-    return
-  }
-  initialized = true
-  initializeMacroProfileRealtimeInfra()
-}
