+++ src/modules/diet/macro-profile/infrastructure/supabase/constants.ts
@@ -1,0 +1,1 @@
+export const SUPABASE_TABLE_MACRO_PROFILES = 'macro_profiles'
