--- src/modules/diet/recipe/application/realtime.ts
@@ -1,10 +1,0 @@
-import { initializeRecipeRealtime as initializeRecipeRealtimeInfra } from '~/modules/diet/recipe/infrastructure/supabase/realtime'
-
-let initialized = false
-export function initializeRecipeRealtime() {
-  if (initialized) {
-    return
-  }
-  initialized = true
-  initializeRecipeRealtimeInfra()
-}
