+++ src/modules/diet/recipe/infrastructure/supabase/constants.ts
@@ -1,0 +1,1 @@
+export const SUPABASE_TABLE_RECIPES = 'recipes'
