+++ src/modules/measure/infrastructure/supabase/constants.ts
@@ -1,0 +1,1 @@
+export const SUPABASE_TABLE_BODY_MEASURES = 'body_measures'
