--- src/modules/search/application/realtime.ts
@@ -1,10 +1,0 @@
-import { initializeCachedSearchRealtime as initializeCachedSearchRealtimeInfra } from '~/modules/search/infrastructure/supabase/realtime'
-
-let initialized = false
-export function initializeCachedSearchRealtime() {
-  if (initialized) {
-    return
-  }
-  initialized = true
-  initializeCachedSearchRealtimeInfra()
-}
