+++ src/modules/search/infrastructure/supabase/constants.ts
@@ -1,0 +1,1 @@
+export const SUPABASE_TABLE_CACHED_SEARCHES = 'cached_searches'
