--- src/modules/weight/application/realtime.ts
@@ -1,7 +1,0 @@
-let initialized = false
-export function initializeWeightRealtime() {
-  if (initialized) {
-    return
-  }
-  initialized = true
-}
